requirements.txt
autogen[gemini,ipython]
openai
pandas
numpy
scikit-learn
python-dotenv
flaml[automl]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib